+++ R/blapply.R
@@ -22,10 +22,10 @@ NULL
 #' @exportMethod blapply
 #' @rdname blapply
 #' @examples
-#'   use("polmineR")
-#'   bt <- partition("GERMAPARLMINI", date = ".*", regex=TRUE)
-#'   speeches <- as.speeches(bt, s_attribute_date = "date", s_attribute_name = "speaker")
-#'   foo <- blapply(speeches, function(x, ...) slot(x, "cpos"))
+#' use("polmineR")
+#' bt <- partition("GERMAPARLMINI", date = ".*", regex=TRUE)
+#' speeches <- as.speeches(bt, s_attribute_date = "date", s_attribute_name = "speaker")
+#' foo <- blapply(speeches, function(x, ...) slot(x, "cpos"))
 #' @importFrom pbapply pblapply
 setGeneric("blapply", function(x, ...) standardGeneric("blapply"))